refactor(website): scope structured data to the website project
CI typecheck surfaced pre-existing failures in tera-website (unrelated
tera-api SDK drift in libs/vue/tera). Touching the shared
vue-unhead-composables lib pulled the broken tera-website into the affected
set, so move the new composable into the website project as a local
composable and limit the change to projects/website, which typechecks clean.

The multi-site rollout (tera/espace/robust + shared legal views) can follow
once the tera-website typecheck debt is resolved.
