+++ libs/vue/unhead/composables/package.json
@@ -2,7 +2,6 @@
   "name": "@lychen/vue-unhead-composables",
   "dependencies": {
     "@lychen/vue-i18n": "workspace:*",
-    "@unhead/schema-org": "^3.1.4",
     "@unhead/vue": "^3.1.4",
     "vue": "^3.5.38",
     "vue-router": "^5.1.0"
