+++ projects/espace/website/src/views/home/View.vue
@@ -173,12 +173,10 @@ import planetMars from './assets/planet-mars.png';
 import { Popover, PopoverContent, PopoverTrigger } from '@lychen/vue-components-core/popover';
 import { LINK } from '@lychen/typescript-espace/constants/App';
 import { useExtendedHead } from '@lychen/vue-unhead-composables/useExtendedHead';
-import { useWebPageSchema } from '@lychen/vue-unhead-composables/useWebPageSchema';
 
 const { t } = usePrefixedI18n(CONFIG);
 
 useExtendedHead(t);
-useWebPageSchema(t);
 
 const planets = [
   { id: 'earth', img: planetEarth },
