+++ projects/robust/website/src/pages/home/PageHome.vue
@@ -16,7 +16,6 @@ import { useI18nExtended } from '@lychen/vue-i18n/composables/useI18nExtended';
 import { messages, TRANSLATION_KEY } from './i18n';
 import PageHomeHero from './PageHomeHero.vue';
 import { useExtendedHead } from '@lychen/vue-unhead-composables/useExtendedHead';
-import { useWebPageSchema } from '@lychen/vue-unhead-composables/useWebPageSchema';
 import Container from '@lychen/vue-components-website/container/Container.vue';
 import Title from '@lychen/vue-components-website/title/Title.vue';
 import PageHomeIntro from './PageHomeIntro.vue';
@@ -27,7 +26,6 @@ import PageHomePeople from './PageHomePeople.vue';
 const { t } = useI18nExtended({ messages, rootKey: TRANSLATION_KEY, prefixed: true });
 
 useExtendedHead(t, { ogImage: ogImageUrl });
-useWebPageSchema(t);
 </script>
 
 <style lang="css" scoped></style>
