+++ projects/tera/website/src/pages/home/PageHome.vue
@@ -10,7 +10,6 @@ import { useI18nExtended } from '@lychen/vue-i18n/composables/useI18nExtended';
 import { messages, TRANSLATION_KEY } from './i18n';
 import PageHomeHero from './PageHomeHero.vue';
 import { useExtendedHead } from '@lychen/vue-unhead-composables/useExtendedHead';
-import { useWebPageSchema } from '@lychen/vue-unhead-composables/useWebPageSchema';
 import SectionApplicationFeatures from '@lychen/vue-components-website/section-application-features/SectionApplicationFeatures.vue';
 import { useFeatures } from '@lychen/vue-tera/composables/use-features/useFeatures';
 import { ref } from 'vue';
@@ -18,7 +17,6 @@ import { ref } from 'vue';
 const { t } = useI18nExtended({ messages, rootKey: TRANSLATION_KEY, prefixed: true });
 
 useExtendedHead(t, { ogImage: ogImageUrl });
-useWebPageSchema(t);
 
 const { getOrganizedFeatures } = useFeatures();
 const features = ref();
