+++ projects/website/src/views/home/View.vue
@@ -11,7 +11,7 @@ import ogImageUrl from './assets/HomeOgImage.webp';
 import { usePrefixedI18n } from '@lychen/vue-i18n/composables/useI18nExtended';
 import { CONFIG } from './i18n';
 import { useExtendedHead } from '@lychen/vue-unhead-composables/useExtendedHead';
-import { useWebPageSchema } from '@lychen/vue-unhead-composables/useWebPageSchema';
+import { useWebPageSchema } from '@/composables/useWebPageSchema';
 import PageHomeHero from './PageHomeHero.vue';
 import PageHomeSymbiosis from './PageHomeSymbiosis.vue';
 import PageHomeODD from './PageHomeODD.vue';
